+++ pkg/gateway/db/postgres/migrations/20210630114505_create_transfer_table.up.sql
@@ -2,6 +2,6 @@ CREATE TABLE IF NOT EXISTS transfer (
    transfer_id SERIAL PRIMARY KEY,
    origin_account SERIAL REFERENCES account (account_id),
    destination_account SERIAL REFERENCES account (account_id),
-   amount DOUBLE PRECISION NOT NULL,
+   amount INTEGER NOT NULL,
    created_at TIMESTAMP WITH TIME ZONE NOT NULL
 );
